move unit tests, and process template vars in clickthroughurl
